Created a separate Table component which was required for my ticket and created some files
