+++ src/components/CustomerProfile/CustomerProfile.jsx
@@ -1,0 +1,10 @@
+
+
+function CustomerProfile(props) {
+    return<tr>
+            `Customer profile ${props.id}`
+        </tr>
+
+}
+
+export default CustomerProfile
