+++ src/components/SearchResult/SearchResult.jsx
@@ -1,40 +1,14 @@
 import "./SearchResult.scss";
-import FakeBookings from "@/data/fakeBookings.json";
-import TableBody from "./TableBody";
-import TableHead from "./TableHead";
-import dayjs from "dayjs";
+import CustomerProfile from "../CustomerProfile/CustomerProfile";
+import Table from "./Table";
 
-const Bookings = FakeBookings;
+
 function SearchResult() {
   return (
-    <table>
-      <thead>
-        <TableHead />
-      </thead>
-      <tbody>
-        {Bookings.map((book) => {
-          const stayNightsTotal = dayjs(book.checkOutDate).diff(
-            dayjs(book.checkInDate),
-            "day"
-          );
-          return (
-              <TableBody
-                key={book.id}
-                id={book.id}
-                title={book.title}
-                firstName={book.firstName}
-                surName={book.surname}
-                email={book.email}
-                roomId={book.roomId}
-                checkInDate={book.checkInDate}
-                checkOutDate={book.checkOutDate}
-                stayNights={stayNightsTotal}
-              />
-          );
-        })}
-      </tbody>
-    </table>
+    <>
+      <Table />
+      <CustomerProfile />
+    </>
   );
 }
 export default SearchResult;
-
