+++ src/components/SearchResult/TableHead.jsx
@@ -11,6 +11,7 @@ function TableHead() {
         <th>Check in date</th>
         <th>Check out date</th>
         <th>Total Nights to Stay</th>
+        <th>Profile</th>
       </tr>
     </>
   );
